lmdb: convert main methods to async
